+++ custom_components/recalbox/api.py
@@ -9,14 +9,14 @@ class RecalboxAPI:
     def __init__(self,
                  host: str = "recalbox.local",
                  api_port_os: int = 80,
-                 api_port_webmanager: int = 81,
+                 api_port_gamesmanager: int = 81,
                  udp_recalbox: int = 1337,
                  udp_emulstation: int = 55355
                  ):
         self.host = host
-        self.api_port_os = api_port_os
-        self.api_port_webmanager = api_port_webmanager
-        self.udp_recalbox = udp_recalbox
+        self.api_port_os = api_port_os # Arrêter, Reboot de Recalbox...
+        self.api_port_gamesmanager = api_port_gamesmanager # Lister les roms, demander un screenshot...
+        self.udp_recalbox = udp_recalbox # Lancer une ROM
         self.udp_emulstation = udp_emulstation
 
     async def send_udp_command(self, port, message):