+++ custom_components/recalbox/config_flow.py
@@ -8,7 +8,7 @@ from .const import DOMAIN
 DATA_SCHEMA = vol.Schema({
     vol.Required("host", default="recalbox.local"): str,
     vol.Required("api_port_os", default=80): int,
-    vol.Required("api_port_webmanager", default=81): int,
+    vol.Required("api_port_gamesmanager", default=81): int,
     vol.Required("udp_recalbox", default=1337): int,
     vol.Required("udp_emulstation", default=55355): int,
 })
@@ -27,6 +27,9 @@ class RecalboxConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         return self.async_show_form(
             step_id="user",
             data_schema=DATA_SCHEMA,
+            sections={
+                "advanced_settings": ("api_port_os", "api_port_gamesmanager", "udp_recalbox", "udp_emulstation")
+            },
             errors=errors,
         )
 
@@ -50,8 +53,11 @@ class RecalboxOptionsFlowHandler(config_entries.OptionsFlow):
             data_schema=vol.Schema({
                 vol.Required("host", default=current_config.get("host", "recalbox.local")): str,
                 vol.Required("api_port_os", default=current_config.get("api_port_os", 80)): int,
-                vol.Required("api_port_webmanager", default=current_config.get("api_port_webmanager", 81)): int,
+                vol.Required("api_port_gamesmanager", default=current_config.get("api_port_gamesmanager", 81)): int,
                 vol.Required("udp_recalbox", default=current_config.get("udp_recalbox", 1337)): int,
                 vol.Required("udp_emulstation", default=current_config.get("udp_emulstation", 55355)): int,
             }),
+            sections={
+                "advanced_settings": ("api_port_os", "api_port_gamesmanager", "udp_recalbox", "udp_emulstation")
+            },
         )